CHANGE OF PACE
I want this to be the simplest project. I originally wanted this to be really unique but with all of the ideas from other projects i decided not to merge anything from past projects into this one. im still keeping the google database features as i really want a multiplayer feel but its going to be more straightforward with my ideas...